Update Fedora version in workflows and Dockerfile
The Fedora version has been updated from 36 to 37 in the GitHub workflows and Dockerfile. This includes changes in the build-deps.yml and build.yml workflow files, a name change for the Fedora Dockerfile, and an update to the Fedora version url within the Dockerfile itself, as well as in the Constants.cs utility file.